dfu: Build warning fixes for 64-bit
Explicitly cast the result of a pointer arithmetic to unsigned int so
that it matches the corresponding printf format string. While at it, use
%p to print a buffer address rather than %x and an explicit cast (which
causes a warning in this case because it's cast to unsigned int instead
of unsigned long).

Cc: Łukasz Majewski <l.majewski@samsung.com>

+++ drivers/dfu/dfu.c
@@ -200,7 +200,7 @@ int dfu_write(struct dfu_entity *dfu, void *buf, int size, int blk_seq_num)
 
 	debug("%s: name: %s buf: 0x%p size: 0x%x p_num: 0x%x offset: 0x%llx bufoffset: 0x%x\n",
 	      __func__, dfu->name, buf, size, blk_seq_num, dfu->offset,
-	      dfu->i_buf - dfu->i_buf_start);
+	      (unsigned int)(dfu->i_buf - dfu->i_buf_start));
 
 	if (!dfu->inited) {
 		/* initial state */
